requirements.txt
appdirs==1.4.4
Appium-Python-Client==1.0.1
attrs==19.3.0
black==19.10b0
click==7.1.2
importlib-metadata==1.6.1
more-itertools==8.3.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
regex==2020.6.8
selenium==3.141.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
